Add drv_pulseclock

+++ src/driver/drv_pulseclock.c
@@ -91,7 +91,7 @@ void PulseClock_onEverySec() {
         phys_daysec += phys_resolution;
         phys_daysec=DaysecNormalise(phys_daysec);
         addLogAdv(LOG_INFO, LOG_FEATURE_DRV, "PulseClock: save to flash %i", phys_daysec);
-        HAL_FlashVars_SaveChannel(PHYS_DAYSEC_FV, phys_daysec);
+        HAL_FlashVars_SaveChannel(PHYS_DAYSEC_FV, (int)phys_daysec);
     }
     else
     {
@@ -137,7 +137,7 @@ void PulseClock_init() {
 	phys_resolution = Tokenizer_GetArgIntegerDefault(1, 60);
 	phys_pulseoffset = Tokenizer_GetArgIntegerDefault(2, 0);
 	phys_pulsemillis = Tokenizer_GetArgIntegerDefault(3, 500);
-	phys_daysec=HAL_FlashVars_GetChannelValue(PHYS_DAYSEC_FV);
+	phys_daysec=(int32_t) HAL_FlashVars_GetChannelValue(PHYS_DAYSEC_FV);
     addLogAdv(LOG_INFO, LOG_FEATURE_DRV, "PulseClock: read from flash %i", phys_daysec);
     if (phys_daysec != DaysecNormalise(phys_daysec))
     {
